feat: refactor model configuration handling and introduce ModelConfig class
