+++ redash/handlers/model_configs.py
@@ -1,18 +1,15 @@
-from typing import List
-
 from flask import request
 
 from redash import models
-from redash.handlers.base import BaseResource, get_object_or_404, require_fields
-from redash.models.models import Model, ModelConfig
+from redash.handlers.base import BaseResource, get_object_or_404, require_fields, abort
+from redash.models.models import Model
+from redash.models.model_config import ModelConfig
 from redash.permissions import require_admin_or_owner, require_permission
-from redash.plywood.objects.data_cube import DataCube
 from redash.serializers.model_serializer import ModelConfigSerializer
 from redash.services.model_config_validator import ModelConfigValidator
 
 UPDATE_ACTION = "update"
 CREATE_ACTION = "create"
-
 
 class ModelsConfigResource(BaseResource):
     @require_permission("edit_model_config")
@@ -44,29 +41,11 @@ class ModelsConfigResource(BaseResource):
 
     @require_permission("view_model_config")
     def get(self, model_id):
-        model: Model = get_object_or_404(Model.get_by_id, model_id)
-        models: list[Model] = Model.query.filter(Model.data_source_id == model.data_source_id).all()
-        data_cubes: List[DataCube.data_cube] = []
-        table_names: list[str] = []
-        cluster_names: set[str] = set()
-
-        for model in models:
-            cube = DataCube(model).data_cube
-            if cube["name"] not in table_names:
-                table_names.append(cube["name"])
-                data_cubes.append(cube)
-                # Collect cluster names from data cubes
-                if "clusterName" in cube:
-                    cluster_names.add(cube["clusterName"])
-
-        # Build cluster objects from collected cluster names
-        clusters = []
-        for cluster_name in cluster_names:
-            if cluster_name != "native":  # native is a special case, doesn't need a cluster object
-                clusters.append({"name": cluster_name, "type": cluster_name})
-
-        return {"appSettings": {"dataCubes": data_cubes, "clusters": clusters, "customization": {}}, "timekeeper": {}}
-
+        config = ModelConfig.get_model_config(model_id)
+        if not config:
+            abort(404)
+        self.record_event({"action": "view", "object_id": model_id, "object_type": "model_config"})
+        return config
 
 class ModelsConfigGetResource(BaseResource):
     @require_permission("view_model_config")
