+++ redash/serializers/__init__.py
@@ -14,6 +14,7 @@ from redash import models
 from redash.models.parameterized_query import ParameterizedQuery
 from redash.permissions import has_access, view_only
 from redash.plywood.objects.data_cube import DataCube
+from redash.models.model_config import ModelConfig
 from redash.serializers.query_result import (
     serialize_query_result,
     serialize_query_result_to_dsv,
@@ -33,7 +34,7 @@ def get_data_cube(model):
     data_cube = DataCube(model=model)
     return data_cube
 
-
+# TODO use redash.plywood.hash_manager.ReportHash instead
 def hash_report(report, can_edit=False):
     # carry this into serializers folder and name it into serialize_report
     data_cube = get_data_cube(report.model)
@@ -69,11 +70,7 @@ def hash_report(report, can_edit=False):
         "is_favorite": is_favorite,
         "is_archived": report.is_archived,
         "landed": True,
-        "appSettings": {
-            "dataCubes": [data_cube.data_cube],
-            "customization": {},
-            "clusters": [],
-        },
+        "appSettings": ModelConfig.get_model_config(report.model_id),
         "id": report.id,
         "api_key": api_key,
         "public_url": public_url,
