+++ redash/services/model_config_validator.py
@@ -1,10 +1,8 @@
 import io
-
 import yaml
 from cerberus import Validator
 from flask_restful import abort
-
-from redash.models.models import ModelConfig
+from redash.models.model_config import ModelConfig
 
 schema = {
     "dataCubes": {
