+++ tests/factories.py
@@ -6,6 +6,7 @@ from redash.models import db
 from redash.permissions import ACCESS_TYPE_MODIFY
 from redash.utils import gen_query_hash, utcnow
 from redash.utils.configuration import ConfigurationContainer
+from redash.models.model_config import ModelConfig
 
 
 class ModelFactory:
@@ -173,7 +174,7 @@ query_snippet_factory = ModelFactory(
 )
 
 model_config_factory = ModelFactory(
-    redash.models.models.ModelConfig,
+    ModelConfig,
     user=user_factory.create,
     content=Sequence("key: {}"),
 )
