+++ tests/models/test_model_configs.py
@@ -1,4 +1,4 @@
-from redash.models.models import ModelConfig
+from redash.models.model_config import ModelConfig
 from tests import BaseTestCase
